quantum-rag/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
sqlalchemy==2.0.35
langchain==0.3.7
langchain-openai==0.2.6
langchain-community==0.3.5
faiss-cpu==1.9.0
openpyxl==3.1.5
pypdf==5.0.1
python-multipart==0.0.12
aiofiles==24.1.0
httpx==0.27.2
